// frontend/src/types.ts
export interface PokemonData {
    attack: number;
    defense: number;
    generation: number;
    hit_points: number;
    legendary: boolean;
    name: string;
    full_name: string;
    number: number;
    special_attack: number;
    special_defense: number;
    speed: number;
    total: number;
    type_one: string;
    type_two: string;
}